Move shard allocator to server common
