+++ core/server-ng/src/server_error.rs
@@ -18,7 +18,7 @@
 use metadata::impls::recovery::RecoveryError;
 // TODO: decouple logging errors from the `server` crate.
 use server::server_error::LogError;
-use server::shard_allocator::ShardingError;
+use server_common::shard_allocator::ShardingError;
 use shard::ShardCtorError;
 use thiserror::Error;
