+++ core/server_common/src/lib.rs
@@ -43,6 +43,7 @@ mod messages_batch_mut;
 mod messages_batch_set;
 mod segment_storage;
 pub mod send_messages2;
+pub mod shard_allocator;
 pub mod sharding;
 
 pub use bootstrap::create_directories;
